# tags: Aggressively inline tag serialization small tag sets