--- a/tags.go
+++ b/tags.go
@@ -76,6 +76,9 @@ func serializeTags(name string, tags map[string]string) string {
 		if t0.dimension > t2.dimension {
 			t0, t2 = t2, t0
 		}
+		if t1.dimension > t3.dimension {
+			t1, t3 = t3, t1
+		}
 		if t1.dimension > t2.dimension {
 			t1, t2 = t2, t1
 		}