--- a/tags_test.go
+++ b/tags_test.go
@@ -85,3 +85,30 @@ func BenchmarkSerializeTags_Two(b *testing.B) {
 		serializeTags(name, tags)
 	}
 }
+
+func BenchmarkSerializeTags_Three(b *testing.B) {
+	const name = "prefix"
+	tags := map[string]string{
+		"tag1": "val1",
+		"tag2": "val2",
+		"tag3": "val3",
+	}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		serializeTags(name, tags)
+	}
+}
+
+func BenchmarkSerializeTags_Four(b *testing.B) {
+	const name = "prefix"
+	tags := map[string]string{
+		"tag1": "val1",
+		"tag2": "val2",
+		"tag3": "val3",
+		"tag4": "val4",
+	}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		serializeTags(name, tags)
+	}
+}