--- a/tags_test.go
+++ b/tags_test.go
@@ -1,10 +1,15 @@
 package stats
 
 import (
+	"bufio"
 	"bytes"
+	crand "crypto/rand"
+	"encoding/base64"
 	"fmt"
+	"math/rand"
 	"sort"
 	"testing"
+	"time"
 )
 
 // Reference serializeTags implementation
@@ -59,6 +64,58 @@ func TestSerializeTagsReference(t *testing.T) {
 		if serialized != expected {
 			t.Errorf("%d Serialized output (%s) didn't match expected output: %s",
 				i, serialized, expected)
+		}
+	}
+}
+
+// Test the network sort used when we have 4 or less tags.  Since the iteration
+// order of maps is random we use random keys in an attempt to get 100% test
+// coverage.
+func TestSerializeTagsNetworkSort(t *testing.T) {
+	const name = "prefix"
+
+	rand.Seed(time.Now().UnixNano())
+	buf := bufio.NewReader(crand.Reader)
+	seen := make(map[string]bool)
+
+	randomString := func() string {
+		for i := 0; i < 100; i++ {
+			b := make([]byte, rand.Intn(30)+1)
+			if _, err := buf.Read(b); err != nil {
+				t.Fatal(err)
+			}
+			s := base64.StdEncoding.EncodeToString(b)
+			if !seen[s] {
+				seen[s] = true
+				return s
+			}
+		}
+		t.Fatal("Failed to generate a random string")
+		return ""
+	}
+
+	makeTags := func(n int) map[string]string {
+		m := make(map[string]string, n)
+		for i := 0; i < n; i++ {
+			k := randomString()
+			v := randomString()
+			m[k] = v
+		}
+		return m
+	}
+
+	// we use a network sort when tag length is 4 or less, but test up to 8
+	// here in case that value is ever increased.
+	for i := 1; i <= 4; i++ {
+		// loop to increase the odds of 100% test coverage
+		for i := 0; i < 10; i++ {
+			tags := makeTags(i)
+			expected := serializeTagsReference(name, tags)
+			serialized := serializeTags(name, tags)
+			if serialized != expected {
+				t.Errorf("%d Serialized output (%s) didn't match expected output: %s",
+					i, serialized, expected)
+			}
 		}
 	}
 }