Most tag sets are small.  This improves runtime by roughly 25% and reduces allocs by roughly 70% when using 4 or less tags.

Benchmark results:

```
benchmark                               old ns/op     new ns/op     delta
BenchmarkSerializeTags_Four-8           341           262           -23.17%

benchmark                               old allocs     new allocs     delta
BenchmarkSerializeTags_Four-8           3              1              -66.67%

benchmark                               old bytes     new bytes     delta
BenchmarkSerializeTags_Four-8           224           64            -71.43%
```