@charlievieth 2019-03-05
@btc this is probably as far as we'd want to go optimizing the `serializeTags()` function.  This is mostly to reduce allocs.  The next big hog is `mergeTags()` that'll be a little more difficult to fix/improve.