@btc 2019-03-05
@charlievieth can you provide benchmarks for tag sets of sizes 1 through 10?