@btc 2019-03-05
can you provide explanatory comments of intent? i think optimizations like this warrant the expression of intent as it's difficult to discern surprising choices that are intentional versus accidental errors.