@charlievieth 2019-03-08
Travis is flaking on `go get golang.org/x/lint/golint` for no good reason.