@charlievieth 2019-03-14
@btc updated to use `b.Run()` - good call.