@btc 2019-03-15
I was initially skeptical about the cost/benefit of these optimizations. However, I checked out the code locally and tried to simplify the code in various ways, observing the effects on performance.

I ultimately concluded that this is about as good as it gets.